move pop utils to own file
